Move createApplePayMethod to wpcom-checkout package
