+++ packages/wpcom-checkout/src/index.ts
@@ -5,3 +5,4 @@ import useDisplayCartMessages from './use-display-cart-messages';
 
 export * from './transformations';
 export { useDisplayCartMessages };
+export { createApplePayMethod } from './payment-methods/apple-pay';
